requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
Pillow>=10.0.0
requests>=2.31.0
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.8.0
numpy<2.0.0
python-dotenv>=1.0.0 